requirements.txt
Flask==3.0.0
huggingface-hub==0.17.3
mysql-connector-python==8.2.0
regex==2023.10.3
requests==2.31.0
tokenizers==0.10.3
torch==1.9.0+cpu
torchaudio==0.9.0
torchvision==0.10.0+cpu
tqdm==4.66.1
transformers==4.10.0
